requirements.txt
alembic==1.14.1
bcrypt==4.3.0
bleach==6.1.0
blinker==1.9.0
boolean.py==4.0
Brotli==1.1.0
CacheControl==0.14.2
cachelib==0.13.0
cairocffi==1.7.1
CairoSVG==2.7.1
certifi==2025.1.31
cffi==1.16.0
charset-normalizer==3.4.1
click==8.1.7
cryptography==44.0.1
cssselect2==0.7.0
cyclonedx-python-lib==8.9.0
defusedxml==0.7.1
Deprecated==1.2.18
et-xmlfile==1.1.0
exceptiongroup==1.2.1
filelock==3.17.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Caching==2.3.1
Flask-Limiter==3.10.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fonttools==4.53.0
greenlet==3.0.3
html5lib==1.1
idna==3.7
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
license-expression==30.4.1
limits==4.0.1
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.20.1
mdurl==0.1.2
msgpack==1.1.0
mysqlclient==2.2.4
openpyxl==3.1.5
ordered-set==4.1.0
packageurl-python==0.16.0
packaging==24.2
pillow==10.3.0
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.8.0
platformdirs==4.3.6
pluggy==1.5.0
py-serializable==1.1.2
pycparser==2.22
pydyf==0.11.0
Pygments==2.19.1
PyMySQL==1.1.1
pyotp==2.9.0
pyparsing==3.2.1
pyphen==0.15.0
pytest==8.3.4
python-dotenv==1.0.1
qrcode==8.0
requests==2.32.3
rich==13.9.4
sentry-sdk==2.22.0
six==1.17.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.30
tinycss2==1.4.0
tinyhtml5==2.0.0
toml==0.10.2
tomli==2.0.1
typing_extensions==4.12.2
urllib3==2.2.3
weasyprint==64.0
webencodings==0.5.1
Werkzeug==3.1.3
wrapt==1.17.2
WTForms==3.2.1
WTForms-SQLAlchemy==0.4.2
zopfli==0.2.3
